% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brush_inputs}
\alias{brush_inputs}
\title{libmypaint brush input metadata}
\usage{
brush_inputs()
}
\description{
libmypaint brush input metadata
}
\examples{
head(brush_inputs())
}
\seealso{
Other brush management:
\code{\link[=brush_dirs]{brush_dirs()}},
\code{\link[=brush_settings]{brush_settings()}},
\code{\link[=brushes]{brushes()}},
\code{\link[=load_brush]{load_brush()}},
\code{\link[=set_brush]{set_brush()}},
\code{\link[=tweak_brush]{tweak_brush()}}
}
\concept{brush management}
