% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brush_presets}
\alias{brush_presets}
\title{Built-in brush presets}
\usage{
brush_presets()
}
\description{
Built-in brush presets
}
\seealso{
Other brush management: 
\code{\link{brush_dirs}()},
\code{\link{brush_inputs}()},
\code{\link{brush_settings}()},
\code{\link{brushes}()},
\code{\link{load_brush}()},
\code{\link{set_brush}()}
}
\concept{brush management}
