% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brush_settings}
\alias{brush_settings}
\title{libmypaint brush setting metadata}
\usage{
brush_settings()
}
\description{
libmypaint brush setting metadata
}
\examples{
head(brush_settings())
}
\seealso{
Other brush management:
\code{\link[=brush_dirs]{brush_dirs()}},
\code{\link[=brush_inputs]{brush_inputs()}},
\code{\link[=brushes]{brushes()}},
\code{\link[=load_brush]{load_brush()}},
\code{\link[=set_brush]{set_brush()}},
\code{\link[=tweak_brush]{tweak_brush()}}
}
\concept{brush management}
