% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rough.R
\name{draw_rough_lines}
\alias{draw_rough_lines}
\title{Draw rough connected lines}
\usage{
draw_rough_lines(x, y = NULL, hand = NULL, ...)
}
\arguments{
\item{x, y}{Coordinates as for \code{\link[graphics:lines]{graphics::lines()}}.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}

\item{...}{Graphics parameters passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
Draws on the current device and returns \code{NULL} invisibly.
}
\description{
Draw rough connected lines
}
\examples{
y <- c(2, 5, 4, 7, 6, 8)
plot(1:6, y, type = "n")
draw_rough_lines(1:6, y, hand = hand(multi_stroke = 2), lwd = 2)
}
\seealso{
Other rough drawing helpers: 
\code{\link{draw_rough_arrows}()},
\code{\link{draw_rough_points}()},
\code{\link{draw_rough_polygons}()},
\code{\link{draw_rough_polypath}()},
\code{\link{draw_rough_rect}()},
\code{\link{draw_rough_segments}()}
}
\concept{rough drawing helpers}
