% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{draw_rough_segments}
\alias{draw_rough_segments}
\title{Draw rough segments}
\usage{
draw_rough_segments(x0, y0, x1, y1, hand = hand(), ...)
}
\arguments{
\item{x0, y0}{Segment starts.}

\item{x1, y1}{Segment ends.}

\item{hand}{Hand-drawn geometry settings created with [hand()].}

\item{...}{Graphics parameters passed to [graphics::lines()].}
}
\value{
Draws on the current device and returns `NULL` invisibly.
}
\description{
Draw rough segments
}
\examples{
plot(1:10, 1:10, type = "n")
draw_rough_segments(1:3, 1:3, 2:4, 3:1)
}
