% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{mypaint_brush_dirs}
\alias{mypaint_brush_dirs}
\title{Discover installed mypaint brush directories}
\usage{
mypaint_brush_dirs()
}
\value{
A character vector of directories containing `.myb` brushes.
}
\description{
Discover installed mypaint brush directories
}
\examples{
mypaint_brush_dirs()
}
