% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{mypaint_brushes}
\alias{mypaint_brushes}
\title{List installed mypaint brushes}
\usage{
mypaint_brushes(paths = default_mypaint_brush_dirs())
}
\arguments{
\item{paths}{Optional brush directories. Defaults to locally discovered
`mypaint-brushes` locations.}
}
\value{
A character vector of brush names, relative to the brush root.
}
\description{
List installed mypaint brushes
}
\examples{
head(mypaint_brushes())
}
