% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{mypaint_load_brush}
\alias{mypaint_load_brush}
\title{Load an installed mypaint brush}
\usage{
mypaint_load_brush(brush, paths = default_mypaint_brush_dirs())
}
\arguments{
\item{brush}{Brush name like `"classic/pencil"` or a path to a `.myb` file.}

\item{paths}{Optional brush directories. Defaults to locally discovered
`mypaint-brushes` locations.}
}
\value{
A JSON brush string suitable for `mypaint_device(brush = ...)`.
}
\description{
Load an installed mypaint brush
}
\examples{
if (length(mypaint_brushes())) {
  x <- mypaint_load_brush(mypaint_brushes()[[1]])
  stopifnot(is.character(x), length(x) == 1L)
}
}
