% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-docs.R
\name{mypaintr-brush-param}
\alias{mypaintr-brush-param}
\title{Brush parameter}
\arguments{
\item{brush}{Brush preset, installed brush name, JSON brush string, named
settings, or \code{NULL} to switch the selected type back to solid rendering.}
}
\description{
Brush parameter
}
\keyword{internal}
