% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-docs.R
\name{mypaintr-brush-paths-param}
\alias{mypaintr-brush-paths-param}
\title{Brush paths parameter}
\arguments{
\item{paths}{Optional brush directories. Defaults to locally discovered
\code{mypaint-brushes} locations.}
}
\description{
Brush paths parameter
}
\keyword{internal}
