% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-docs.R
\name{mypaintr-brush-settings-param}
\alias{mypaintr-brush-settings-param}
\title{Brush settings parameter}
\arguments{
\item{settings}{Named settings overriding \code{brush}.}
}
\description{
Brush settings parameter
}
\keyword{internal}
