% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-docs.R
\name{mypaintr-rough-fill}
\alias{mypaintr-rough-fill}
\title{Rough drawing fill parameters}
\arguments{
\item{col}{Fill colour. When visible and \code{fill_pattern} is \code{NULL}, a solid
fill is drawn.}

\item{border}{Border colour.}

\item{fill_pattern}{Optional fill pattern created with \code{\link[=hatch]{hatch()}},
\code{\link[=crosshatch]{crosshatch()}}, \code{\link[=zigzag]{zigzag()}}, or \code{\link[=jumble]{jumble()}}.}
}
\description{
Rough drawing fill parameters
}
\keyword{internal}
