% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-docs.R
\name{mypaintr-rough-hand}
\alias{mypaintr-rough-hand}
\title{Rough drawing hand parameter}
\arguments{
\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}
}
\description{
Rough drawing hand parameter
}
\keyword{internal}
