% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hands.R
\name{set_hand}
\alias{set_hand}
\title{Set the active hand}
\usage{
set_hand(hand = NULL, type = c("both", "stroke", "fill"))
}
\arguments{
\item{hand}{Hand-drawn geometry created with \code{\link[=hand]{hand()}}, or \code{NULL} to disable
it for the selected type. This disables rough path perturbation only; it
does not disable the active brush, and note that some brushes have
their own internal wobbly pathing! Use \code{\link[=set_brush]{set_brush()}} as well if you want
fully plain, solid rendering.}

\item{type}{Which rendering channel to update: \code{"both"}, \code{"stroke"}, or
\code{"fill"}.}
}
\value{
\code{NULL}, invisibly. If the active graphics device is not
\code{\link[=mypaint_device]{mypaint_device()}}, this emits a warning and has no effect.
}
\description{
Set the active hand
}
\examples{
ex_file <- tempfile(fileext = ".png")
mypaint_device(ex_file)

plot.new()
plot.window(c(0, 10), c(0, 10))
set_hand(hand())
rect(1, 1, 5, 5, col = "darkred", density = 5)
set_hand(human_hand())
rect(5, 5, 9, 9, col = "darkgreen", density = 5)

dev.off()
img <- png::readPNG(ex_file)
grid::grid.raster(img)
}
