% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brush_dirs}
\alias{brush_dirs}
\title{Discover installed mypaint brush directories}
\usage{
brush_dirs()
}
\value{
A character vector of directories containing \code{.myb} brushes.
}
\description{
Discover installed mypaint brush directories
}
\examples{
brush_dirs()
}
