% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brush_dirs}
\alias{brush_dirs}
\title{Discover installed mypaint brush directories}
\usage{
brush_dirs()
}
\value{
A character vector of directories containing \code{.myb} brushes.
}
\description{
Discover installed mypaint brush directories
}
\examples{
brush_dirs()
}
\seealso{
Other brush management: 
\code{\link{brush_inputs}()},
\code{\link{brush_settings}()},
\code{\link{brushes}()},
\code{\link{load_brush}()},
\code{\link{set_brush}()},
\code{\link{tweak_brush}()}
}
\concept{brush management}
