% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brush_inputs}
\alias{brush_inputs}
\title{libmypaint brush input metadata}
\usage{
brush_inputs()
}
\description{
libmypaint brush input metadata
}
\examples{
head(brush_inputs())
}
\seealso{
Other brush management: 
\code{\link{brush_dirs}()},
\code{\link{brush_presets}()},
\code{\link{brush_settings}()},
\code{\link{brushes}()},
\code{\link{load_brush}()},
\code{\link{set_brush}()}
}
\concept{brush management}
