% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brush_settings}
\alias{brush_settings}
\title{libmypaint brush setting metadata}
\usage{
brush_settings()
}
\description{
libmypaint brush setting metadata
}
\examples{
head(brush_settings())
}
\seealso{
Other brush management: 
\code{\link{brush_dirs}()},
\code{\link{brush_inputs}()},
\code{\link{brush_presets}()},
\code{\link{brushes}()},
\code{\link{load_brush}()},
\code{\link{set_brush}()}
}
\concept{brush management}
