% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brushes}
\alias{brushes}
\title{List installed mypaint brushes}
\usage{
brushes(paths = default_mypaint_brush_dirs())
}
\arguments{
\item{paths}{Optional brush directories. Defaults to locally discovered
\code{mypaint-brushes} locations.}
}
\value{
A character vector of brush names, relative to the brush root.
}
\description{
List installed mypaint brushes
}
\examples{
head(brushes())
}
