% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{brushes}
\alias{brushes}
\title{List installed mypaint brushes}
\usage{
brushes(paths = default_mypaint_brush_dirs())
}
\arguments{
\item{paths}{Optional brush directories. Defaults to locally discovered
\code{mypaint-brushes} locations.}
}
\value{
A character vector of brush names, relative to the brush root.
}
\description{
List installed mypaint brushes
}
\examples{
head(brushes())
}
\seealso{
Other brush management: 
\code{\link{brush_dirs}()},
\code{\link{brush_inputs}()},
\code{\link{brush_settings}()},
\code{\link{load_brush}()},
\code{\link{set_brush}()},
\code{\link{tweak_brush}()}
}
\concept{brush management}
