% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{draw_rough_lines}
\alias{draw_rough_lines}
\title{Draw rough connected lines}
\usage{
draw_rough_lines(x, y = NULL, hand = NULL, ...)
}
\arguments{
\item{x, y}{Coordinates as for [graphics::lines()].}

\item{hand}{Hand-drawn geometry settings created with [hand()].}

\item{...}{Graphics parameters passed to [graphics::lines()].}
}
\value{
Draws on the current device and returns `NULL` invisibly.
}
\description{
Draw rough connected lines
}
\examples{
plot(1:10, cumsum(rnorm(10)), type = "n")
draw_rough_lines(1:10, cumsum(rnorm(10)))
}
