% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rough.R
\name{draw_rough_points}
\alias{draw_rough_points}
\title{Draw rough points}
\usage{
draw_rough_points(x, y = NULL, hand = NULL, ...)
}
\arguments{
\item{x, y}{Point coordinates as for \code{\link[graphics:points]{graphics::points()}}.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}

\item{...}{Graphics parameters passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\value{
Draws on the current device and returns \code{NULL} invisibly.
}
\description{
Draw rough points
}
\examples{
plot(1:10, 1:10, type = "n")
draw_rough_points(1:10, 1:10, pch = 16)
}
\seealso{
Other rough drawing helpers: 
\code{\link{draw_rough_arrows}()},
\code{\link{draw_rough_lines}()},
\code{\link{draw_rough_polygons}()},
\code{\link{draw_rough_polypath}()},
\code{\link{draw_rough_rect}()},
\code{\link{draw_rough_segments}()}
}
\concept{rough drawing helpers}
