% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rough.R
\name{draw_rough_segments}
\alias{draw_rough_segments}
\title{Draw rough segments}
\usage{
draw_rough_segments(x0, y0, x1, y1, hand = NULL, ...)
}
\arguments{
\item{x0, y0}{Segment starts.}

\item{x1, y1}{Segment ends.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}

\item{...}{Graphics parameters passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
Draws on the current device and returns \code{NULL} invisibly.
}
\description{
Draw rough segments
}
\examples{
plot(1:10, 1:10, type = "n")
draw_rough_segments(1:3, 2:4, 4:6, c(8, 5, 7), lwd = 2)
}
\seealso{
Other rough drawing helpers: 
\code{\link{draw_rough_arrows}()},
\code{\link{draw_rough_lines}()},
\code{\link{draw_rough_points}()},
\code{\link{draw_rough_polygons}()},
\code{\link{draw_rough_polypath}()},
\code{\link{draw_rough_rect}()}
}
\concept{rough drawing helpers}
