% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\docType{package}
\name{mypaintr-package}
\alias{mypaintr}
\alias{mypaintr-package}
\title{mypaintr: Artistic Graphics Device Backed by libmypaint}
\description{
Raster graphics device for base R graphics that renders stroked primitives with libmypaint and uses Cairo for text and solid fills on the same backing surface.
}
\author{
\strong{Maintainer}: David Hugh-Jones \email{david@example.com}

}
\keyword{internal}
