% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device-style.R
\docType{package}
\name{mypaintr-package}
\alias{mypaintr}
\alias{mypaintr-package}
\title{mypaintr: Plot R Graphics Like a Human}
\description{
R graphics device for human-like, sketched plotting. Uses libmypaint brushes to draw lines and polygons. Includes functions for plotting "rough" lines and polygons. Integrates with base and "ggplot" graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hughjonesd/mypaintr}
  \item \url{https://hughjonesd.github.io/mypaintr/}
  \item Report bugs at \url{https://github.com/hughjonesd/mypaintr/issues}
}

}
\author{
\strong{Maintainer}: David Hugh-Jones \email{david@example.com}

}
\keyword{internal}
