% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rough.R
\name{rough_lines}
\alias{rough_lines}
\alias{draw_rough_lines}
\title{Compute or draw rough connected lines}
\usage{
rough_lines(x, y = NULL, hand = NULL)

draw_rough_lines(x, y = NULL, hand = NULL, ...)
}
\arguments{
\item{x, y}{Coordinates as for \code{\link[graphics:lines]{graphics::lines()}}.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}

\item{...}{Graphics parameters passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
A list with \code{x}, \code{y}, and \code{id} components describing roughened
polyline geometry for each connected run.
}
\description{
Compute or draw rough connected lines
}
\examples{
y <- c(2, 5, 4, 7, 6, 8)
plot(1:6, y, type = "n")
draw_rough_lines(1:6, y, hand = human_hand(multi_stroke = 2))
}
\seealso{
Other rough drawing helpers:
\code{\link[=rough_arrows]{rough_arrows()}},
\code{\link[=rough_points]{rough_points()}},
\code{\link[=rough_polygons]{rough_polygons()}},
\code{\link[=rough_polypath]{rough_polypath()}},
\code{\link[=rough_rect]{rough_rect()}},
\code{\link[=rough_segments]{rough_segments()}}
}
\concept{rough drawing helpers}
