% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rough.R
\name{rough_points}
\alias{rough_points}
\alias{draw_rough_points}
\title{Compute or draw rough points}
\usage{
rough_points(x, y = NULL, hand = NULL)

draw_rough_points(x, y = NULL, hand = NULL, ...)
}
\arguments{
\item{x, y}{Point coordinates as for \code{\link[graphics:points]{graphics::points()}}.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}

\item{...}{Graphics parameters passed to \code{\link[graphics:points]{graphics::points()}}.}
}
\value{
A list with jittered \code{x} and \code{y} point locations.
}
\description{
Compute or draw rough points
}
\examples{
plot(1:5, 1:5, type = "n")
rough_points(1:5, 1:5, hand = hand(endpoint_jitter = 0.02))
plot(1:10, 1:10, type = "n")
draw_rough_points(1:10, 1:10, pch = 16, cex = 1.4)
}
\seealso{
Other rough drawing helpers: 
\code{\link{rough_arrows}()},
\code{\link{rough_lines}()},
\code{\link{rough_polygons}()},
\code{\link{rough_polypath}()},
\code{\link{rough_rect}()},
\code{\link{rough_segments}()}
}
\concept{rough drawing helpers}
