% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{rough_polygon}
\alias{rough_polygon}
\title{Compute a rough polygon outline}
\usage{
rough_polygon(x, y = NULL, hand = NULL)
}
\arguments{
\item{x, y}{Polygon coordinates.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}
}
\value{
A list with \code{x} and \code{y} components containing a roughened closed
outline suitable for plotting with \code{\link[graphics:lines]{graphics::lines()}}.
}
\description{
Compute a rough polygon outline
}
\examples{
rough_polygon(c(2, 5, 8, 3), c(2, 7, 5, 1))
}
