% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{rough_rect}
\alias{rough_rect}
\title{Compute a rough rectangle outline}
\usage{
rough_rect(x0, y0, x1, y1, hand = NULL)
}
\arguments{
\item{x0, y0}{Rectangle corner.}

\item{x1, y1}{Opposite rectangle corner.}

\item{hand}{Hand-drawn geometry settings created with [hand()].}
}
\value{
A list with `x` and `y` components containing a roughened closed
  outline suitable for plotting with [graphics::lines()].
}
\description{
Compute a rough rectangle outline
}
\examples{
rough_rect(2, 2, 5, 6)
}
