% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-rough.R
\name{rough_segments}
\alias{rough_segments}
\alias{draw_rough_segments}
\title{Compute or draw rough segments}
\usage{
rough_segments(x0, y0, x1, y1, hand = NULL)

draw_rough_segments(x0, y0, x1, y1, hand = NULL, ...)
}
\arguments{
\item{x0, y0}{Segment starts.}

\item{x1, y1}{Segment ends.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}

\item{...}{Graphics parameters passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\value{
A list with \code{x}, \code{y}, and \code{id} components describing roughened
polyline geometry for each segment.
}
\description{
Compute or draw rough segments
}
\examples{
plot(1:10, 1:10, type = "n")
draw_rough_segments(1:3, 2:4, 4:6, c(8, 5, 7), hand = human_hand())
}
\seealso{
Other rough drawing helpers:
\code{\link[=rough_arrows]{rough_arrows()}},
\code{\link[=rough_lines]{rough_lines()}},
\code{\link[=rough_points]{rough_points()}},
\code{\link[=rough_polygons]{rough_polygons()}},
\code{\link[=rough_polypath]{rough_polypath()}},
\code{\link[=rough_rect]{rough_rect()}}
}
\concept{rough drawing helpers}
