% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{rough_segments}
\alias{rough_segments}
\title{Compute rough segment geometry}
\usage{
rough_segments(x0, y0, x1, y1, hand = NULL)
}
\arguments{
\item{x0, y0}{Segment starts.}

\item{x1, y1}{Segment ends.}

\item{hand}{Hand-drawn geometry settings created with \code{\link[=hand]{hand()}}.}
}
\value{
A list with \code{x}, \code{y}, and \code{id} components describing roughened
polyline geometry for each segment.
}
\description{
Compute rough segment geometry
}
\examples{
rough_segments(1:2, 1:2, 2:3, 3:2)
}
