% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaintr.R
\name{set_hand}
\alias{set_hand}
\title{Set the active mypaintr hand-drawn geometry}
\usage{
set_hand(hand = NULL, type = c("both", "stroke", "fill"))
}
\arguments{
\item{hand}{Hand-drawn geometry created with \code{\link[=hand]{hand()}}, or \code{NULL} to disable
it for the selected type.}

\item{type}{Which rendering channel to update: \code{"both"}, \code{"stroke"}, or
\code{"fill"}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Set the active mypaintr hand-drawn geometry
}
