% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushes.R
\name{tweak_brush}
\alias{tweak_brush}
\title{Create a reusable tweaked brush specification}
\usage{
tweak_brush(brush, ..., normalize = "all")
}
\arguments{
\item{brush}{Installed brush name, \code{.myb} file path, JSON brush string, or
another \code{\link[=tweak_brush]{tweak_brush()}} object.}

\item{...}{Named libmypaint base-value overrides.}

\item{normalize}{One of \code{"all"}, \code{"size"}, \code{"tracking"}, or \code{"none"}.}
}
\value{
A reusable brush specification object.
}
\description{
Create a reusable tweaked brush specification
}
\examples{
if ("classic/pen" \%in\% brushes()) {
  tweak_brush("classic/pen", normalize = "tracking", radius_logarithmic = log(1.2))
}
}
\seealso{
Other brush management: 
\code{\link{brush_dirs}()},
\code{\link{brush_inputs}()},
\code{\link{brush_settings}()},
\code{\link{brushes}()},
\code{\link{load_brush}()},
\code{\link{set_brush}()}
}
\concept{brush management}
